Remove userEntity pattern and change it to only user model
