--- server/src/main/java/com/g13cs3219/server/dto/User.java
@@ -1,17 +1,0 @@
-package com.g13cs3219.server.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class User {
-    private Long userId;
-    private String username;
-    private String email;
-    private String password;
-}
