+++ server/src/main/java/com/g13cs3219/server/model/User.java
@@ -1,0 +1,76 @@
+package com.g13cs3219.server.model;
+
+import java.util.Date;
+
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Entity
+@Table(name = "users")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "user_id_seq")
+    @Column(name = "user_id")
+    private Long userId;
+
+    @Column(nullable = false)
+    private String email;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    @Builder.Default
+    private Role role = Role.USER;
+
+    private String name;
+
+    @Column(nullable = false)
+    private String username;
+
+    @Column(nullable = false, name = "encoded_password")
+    private String encodedPassword;
+
+    @Column(name = "avatar_url")
+    private String avatarUrl;
+
+    @Column(name = "preferred_language")
+    private String preferredLanguage;
+
+    @Column(name = "preferred_topic")
+    private String preferredTopic;
+
+    @Column(name = "created_by")
+    private int createdBy;
+
+    @CreationTimestamp
+    @Column(name = "created_at", updatable = false)
+    private String createdAt;
+
+    @Column(name = "updated_by")
+    private int updatedBy;
+
+    @UpdateTimestamp
+    @Column(name = "updated_at")
+    private Date updatedAt;
+
+    @Column(name = "is_active")
+    @Builder.Default
+    private boolean isActive = true;
+}
